fortune_web_server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
dj-database-url==2.2.0
    # via -r requirements.in
django==5.0.7
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.4.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
packaging==24.1
    # via gunicorn
psycopg2-binary==2.9.9
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
sqlparse==0.5.0
    # via django
typing-extensions==4.12.2
    # via dj-database-url
whitenoise==6.7.0
    # via -r requirements.in
